+++ Frameworks/AVFoundation.framework/AVUnreachableAssetInspectorLoader.h
@@ -5,14 +5,14 @@
 @class NSDictionary;
 
 @interface AVUnreachableAssetInspectorLoader : AVAssetInspectorLoader {
-    int _figError;
+    long _figError;
     NSDictionary *_figErrorUserInfo;
 }
 
 - (Class)_classForTrackInspectors;
 - (void)dealloc;
-- (id)initWithFigError:(int)arg1 userInfo:(id)arg2;
+- (id)initWithFigError:(long)arg1 userInfo:(id)arg2;
 - (void)loadValuesAsynchronouslyForKeys:(id)arg1 completionHandler:(id)arg2;
-- (long long)statusOfValueForKey:(id)arg1 error:(id*)arg2;
+- (int)statusOfValueForKey:(id)arg1 error:(id*)arg2;
 
 @end
