+++ Frameworks/AddressBookUI.framework/ABContactActionSplitCell.h
@@ -15,10 +15,10 @@
 - (void)_setupButton:(id)arg1 forAction:(id)arg2;
 - (id)constantConstraints;
 - (void)dealloc;
-- (id)initWithStyle:(long long)arg1 reuseIdentifier:(id)arg2;
+- (id)initWithStyle:(int)arg1 reuseIdentifier:(id)arg2;
 - (id)labelView;
 - (id)leftLabel;
-- (double)minCellHeight;
+- (float)minCellHeight;
 - (id)rightLabel;
 - (id)rightMostView;
 - (void)setCardGroupItem:(id)arg1;
