+++ Frameworks/AddressBookUI.framework/ABLabelViewWithVariablePositioning.h
@@ -5,30 +5,30 @@
 @class UILabel;
 
 @interface ABLabelViewWithVariablePositioning : UIView {
-    double _bottomPadding;
-    double _horizontalPadding;
+    float _bottomPadding;
+    float _horizontalPadding;
     UILabel *_label;
-    double _topPadding;
+    float _topPadding;
 }
 
-@property double bottomPadding;
-@property double horizontalPadding;
-@property double topPadding;
+@property float bottomPadding;
+@property float horizontalPadding;
+@property float topPadding;
 
-- (double)bottomPadding;
+- (float)bottomPadding;
 - (void)dealloc;
-- (double)horizontalPadding;
-- (id)initWithFrame:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
+- (float)horizontalPadding;
+- (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
 - (void)layoutSubviews;
-- (void)setBottomPadding:(double)arg1;
-- (void)setHorizontalPadding:(double)arg1;
+- (void)setBottomPadding:(float)arg1;
+- (void)setHorizontalPadding:(float)arg1;
 - (void)setLabelFont:(id)arg1;
 - (void)setLabelShadowColor:(id)arg1;
-- (void)setLabelShadowOffset:(struct CGSize { double x1; double x2; })arg1;
+- (void)setLabelShadowOffset:(struct CGSize { float x1; float x2; })arg1;
 - (void)setLabelText:(id)arg1;
 - (void)setLabelTextColor:(id)arg1;
-- (void)setTopPadding:(double)arg1;
-- (struct CGSize { double x1; double x2; })sizeThatFits:(struct CGSize { double x1; double x2; })arg1;
-- (double)topPadding;
+- (void)setTopPadding:(float)arg1;
+- (struct CGSize { float x1; float x2; })sizeThatFits:(struct CGSize { float x1; float x2; })arg1;
+- (float)topPadding;
 
 @end
