+++ Frameworks/AddressBookUI.framework/ABMedicalIDAction.h
@@ -7,21 +7,21 @@
 @interface ABMedicalIDAction : ABPropertyAction <HKMedicalIDViewControllerDelegate> {
     _HKMedicalIDData *_healthData;
     HKHealthStore *_healthStore;
-    long long _medicalIDActionType;
+    int _medicalIDActionType;
 }
 
 @property(retain) CNContact * contact;
 @property(copy,readonly) NSString * debugDescription;
 @property(copy,readonly) NSString * description;
-@property(readonly) unsigned long long hash;
+@property(readonly) unsigned int hash;
 @property(retain) _HKMedicalIDData * healthData;
 @property(retain) HKHealthStore * healthStore;
-@property long long medicalIDActionType;
+@property int medicalIDActionType;
 @property(readonly) Class superclass;
 
 - (id)healthData;
 - (id)healthStore;
-- (long long)medicalIDActionType;
+- (int)medicalIDActionType;
 - (void)medicalIDViewControllerDidCancel:(id)arg1;
 - (void)medicalIDViewControllerDidDelete:(id)arg1;
 - (void)medicalIDViewControllerDidFinish:(id)arg1;
@@ -29,6 +29,6 @@
 - (void)performActionWithSender:(id)arg1;
 - (void)setHealthData:(id)arg1;
 - (void)setHealthStore:(id)arg1;
-- (void)setMedicalIDActionType:(long long)arg1;
+- (void)setMedicalIDActionType:(int)arg1;
 
 @end
