+++ Frameworks/AddressBookUI.framework/ABStarkStyleProvider.h
@@ -5,16 +5,16 @@
 @interface ABStarkStyleProvider : ABStyleProvider {
 }
 
-+ (double)memberRowHeight;
++ (float)memberRowHeight;
 
 - (id)memberNameBoldFont;
 - (id)memberNameDisabledTextColor;
-- (double)memberNameFontSize;
+- (float)memberNameFontSize;
 - (id)memberNamePlaceholderFont;
 - (id)memberNameRegularFont;
 - (id)memberNameTextColor;
 - (id)membersBackgroundColor;
-- (unsigned long long)membersCountToShowIndex;
+- (unsigned int)membersCountToShowIndex;
 - (id)membersSearchBackgroundColor;
 - (id)membersSearchCellBackgroundColor;
 - (id)personLabelHighlightedColor;
