+++ Frameworks/CFNetwork.framework/__NSCFMemoryHTTPCookieStorage.h
@@ -7,14 +7,14 @@
 
 + (void)initialize;
 
-- (unsigned long long)cookieAcceptPolicy;
+- (unsigned int)cookieAcceptPolicy;
 - (id)cookies;
 - (id)cookiesForURL:(id)arg1;
 - (void)deleteCookie:(id)arg1;
 - (id)description;
 - (id)init;
 - (void)setCookie:(id)arg1;
-- (void)setCookieAcceptPolicy:(unsigned long long)arg1;
+- (void)setCookieAcceptPolicy:(unsigned int)arg1;
 - (void)setCookies:(id)arg1 forURL:(id)arg2 mainDocumentURL:(id)arg3;
 - (id)sortedCookiesUsingDescriptors:(id)arg1;
