+++ Frameworks/CFNetwork.framework/__NSCFMemoryURLCache.h
@@ -5,9 +5,9 @@
 @class NSMutableArray, NSMutableDictionary, NSObject<OS_dispatch_queue>;
 
 @interface __NSCFMemoryURLCache : NSObject {
-    unsigned long long _currentLength;
+    unsigned int _currentLength;
     id _internal;
-    unsigned long long _memoryCapacity;
+    unsigned int _memoryCapacity;
     NSObject<OS_dispatch_queue> *_memoryQueue;
     NSMutableDictionary *_memoryStorage;
     NSMutableArray *_memoryStorageLRU;
@@ -16,20 +16,20 @@
 + (void)initialize;
 
 - (id)cachedResponseForRequest:(id)arg1;
-- (unsigned long long)currentDiskUsage;
-- (unsigned long long)currentMemoryUsage;
+- (unsigned int)currentDiskUsage;
+- (unsigned int)currentMemoryUsage;
 - (void)dealloc;
 - (id)description;
-- (unsigned long long)diskCapacity;
+- (unsigned int)diskCapacity;
 - (id)initByCallingSuperSuperInit;
 - (id)initEmptyCache;
 - (id)initMemoryCache;
-- (id)initWithMemoryCapacity:(unsigned long long)arg1 diskCapacity:(unsigned long long)arg2 diskPath:(id)arg3;
-- (unsigned long long)memoryCapacity;
+- (id)initWithMemoryCapacity:(unsigned int)arg1 diskCapacity:(unsigned int)arg2 diskPath:(id)arg3;
+- (unsigned int)memoryCapacity;
 - (void)removeAllCachedResponses;
 - (void)removeCachedResponseForRequest:(id)arg1;
-- (void)setDiskCapacity:(unsigned long long)arg1;
-- (void)setMemoryCapacity:(unsigned long long)arg1;
+- (void)setDiskCapacity:(unsigned int)arg1;
+- (void)setMemoryCapacity:(unsigned int)arg1;
 - (void)storeCachedResponse:(id)arg1 forRequest:(id)arg2;
 
 @end
