+++ Frameworks/CoreAudioKit.framework/CAUITransportButton.h
@@ -5,12 +5,12 @@
 @interface CAUITransportButton : UIButton {
     struct CGRect { 
         struct CGPoint { 
-            double x; 
-            double y; 
+            float x; 
+            float y; 
         } origin; 
         struct CGSize { 
-            double width; 
-            double height; 
+            float width; 
+            float height; 
         } size; 
     int drawingStyle;
     struct CGColor { } *fillColor;
@@ -26,7 +26,7 @@
 - (struct CGColor { }*)fillColor;
 - (void)flash;
 - (id)initWithCoder:(id)arg1;
-- (id)initWithFrame:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
+- (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
 - (struct CGPath { }*)newPathRefForStyle:(int)arg1;
 - (void)setDrawingStyle:(int)arg1;
 - (void)setFillColor:(struct CGColor { }*)arg1;
