+++ Frameworks/CoreAuthentication.framework/LALegacyHelper.h
@@ -7,17 +7,17 @@
 
 + (id)_eventProcessingFromVREventFilter:(id)arg1 error:(id*)arg2;
 + (id)_eventProcessingFromVREventProcessing:(id)arg1 error:(id*)arg2;
-+ (long long)eventFromVREvent:(id)arg1 error:(id*)arg2;
-+ (long long)eventFromVREvent:(id)arg1;
-+ (long long)optionFromVRHint:(id)arg1 error:(id*)arg2;
++ (int)eventFromVREvent:(id)arg1 error:(id*)arg2;
++ (int)eventFromVREvent:(id)arg1;
++ (int)optionFromVRHint:(id)arg1 error:(id*)arg2;
 + (id)optionsFromVRHints:(id)arg1 error:(id*)arg2;
 + (id)optionsFromVRHints:(id)arg1;
-+ (long long)policyFromVRPolicy:(id)arg1 error:(id*)arg2;
-+ (long long)policyFromVRPolicy:(id)arg1;
-+ (long long)resultFromVRResult:(id)arg1 error:(id*)arg2;
-+ (long long)resultFromVRResult:(id)arg1;
++ (int)policyFromVRPolicy:(id)arg1 error:(id*)arg2;
++ (int)policyFromVRPolicy:(id)arg1;
++ (int)resultFromVRResult:(id)arg1 error:(id*)arg2;
++ (int)resultFromVRResult:(id)arg1;
 + (id)uiDelegateFromVRUIDelegate:(id)arg1;
 + (id)vrErrorFromLAError:(id)arg1;
-+ (id)vrErrorWithCode:(long long)arg1 andMessage:(id)arg2;
++ (id)vrErrorWithCode:(int)arg1 andMessage:(id)arg2;
 
 @end
