+++ Frameworks/CoreBluetooth.framework/CBCentral.h
@@ -5,10 +5,10 @@
 @interface CBCentral : CBPeer {
 }
 
-@property(readonly) unsigned long long maximumUpdateValueLength;
+@property(readonly) unsigned int maximumUpdateValueLength;
 
 - (id)description;
 - (id)initWithUUID:(id)arg1;
-- (unsigned long long)maximumUpdateValueLength;
+- (unsigned int)maximumUpdateValueLength;
 
 @end
